+++ src/main/java/org/ow2/authzforce/core/pdp/api/value/StringParseableValue.java
@@ -1,5 +1,5 @@
 /**
- * Copyright 2012-2018 Thales Services SAS.
+ * Copyright 2012-2019 THALES.
  *
  * This file is part of AuthzForce CE.
  *
@@ -29,11 +29,12 @@ import net.sf.saxon.s9api.XPathCompiler;
  * Attribute Value based on string or possibly other type of {@link Serializable} content, without any extra XML attributes.
  *
  * @param <V>
+ *            underlying Java data type
  */
 public abstract class StringParseableValue<V> extends SimpleValue<V>
 {
 
-	protected StringParseableValue(V rawVal) throws IllegalArgumentException, NullPointerException
+	protected StringParseableValue(final V rawVal) throws IllegalArgumentException, NullPointerException
 	{
 		super(rawVal);
 	}